FSW-91 C4018: signed-unsigned-mismatch
* FabianLib: Fix C4018 by casting 'unsigned int' to 'int' prior to using it in computation.

* CELib: Fix C4018 - loop iterators should be unsigned.

* FabianFHO: Fix C4018 for Diagramm.cpp

* FabianHFO: Fix C4018 for remaining issues

+++ FabianHFO/InterfaceAcuLink_SERIAL.cpp
@@ -633,7 +633,7 @@ int CInterfaceAcuLink_SERIAL::setPoint( char *s, char *news, int nk )
 	}
 
 	j = 0;
-	if( strlen( s ) <=  nk )
+	if( strlen( s ) <=  (size_t)nk )
 	{
 		for( i = strlen( s ), j = 0; i < nk + 1; i++ )
 		{
